+++ src/Services/Webauthn/AbstractValidatorFactory.php
@@ -6,7 +6,6 @@ use CBOR\Decoder;
 use Cose\Algorithm\Manager;
 use Cose\Algorithm\Signature;
 use CBOR\Tag\TagObjectManager;
-use Http\Adapter\Guzzle6\Client;
 use Webauthn\PublicKeyCredentialLoader;
 use CBOR\OtherObject\OtherObjectManager;
 use Webauthn\AttestationStatement\AttestationObjectLoader;
@@ -51,7 +50,12 @@ abstract class AbstractValidatorFactory extends AbstractFactory
 
         // https://www.w3.org/TR/webauthn/#android-safetynet-attestation
         if ($this->config->get('webauthn.google_safetynet_api_key', '') !== '') {
-            $attestationStatementSupportManager->add(new AndroidSafetyNetAttestationStatementSupport(new Client(), $this->config->get('webauthn.google_safetynet_api_key')));
+            try {
+                $client = \Http\Discovery\HttpClientDiscovery::find();
+                $attestationStatementSupportManager->add(new AndroidSafetyNetAttestationStatementSupport($client, $this->config->get('webauthn.google_safetynet_api_key')));
+            } catch (\Http\Discovery\Exception\NotFoundException $e) {
+                // ignore
+            }
         }
 
         // https://www.w3.org/TR/webauthn/#android-key-attestation